requirements.txt
torch
torchvision
Pillow
matplotlib
streamlit
